Added Get-WindowsVersionComparison
Detects current Win10 version and compares against another version. Determines if the current windows version is "equal to", "less than", "greater than",
"less than or equal to", or "greather than or equal to"

In the case of an invalid situation, such as requesting a version of windows that is not supported, or running this function on a Windows Server machine,
this script will throw an exception, so it should be used with try/catch.

Upon meeting a valid situation. It will check the version you provided against the version of windows that the current machine is running and output a hash table with "Result" (boolean) and Output (string).

Example of correct usage:
Try {
    $winIsLessThan20H2 = Get-WindowsVersionComparison -LessThan '20H2'
} Catch {
    Write-Output $Error[0].Exception.Message
}

If ($winIsLessThan20H2.Result) {
    Return 'Success! This is a newer version!' + $winIsLessThan20H2.Output
} Else {
    Return 'Oh no! This is an old version!' + $winIsLessThan20H2.Output
}