New function(s) alert -> Registry-Helpers
All functions rely on $regPath being set before this script is called. The intention is for a given script to have a single reg path where it's interacting with keys. This is for the sake of convenience since this scenario pops up often. Each function optionally receives a -Path param which will be used instead if provided, for when this convention doesn't fit.