fixed it
no longer need to put in `()` but you do need to `invoke-expression` to use the code directly, like:
```
$com = New-EncodedCommandWithArguments -ScriptBlock $sb -ArgumentList @(2,2)

powershell.exe -Command ($com | invoke-expression)
```

This is not the case when used for intended purpose inside a string like this:

```
$task = New-ScheduledTaskAction -Execute "PowerShell.exe" -Argument "-NoProfile -Command $(New-EncodedCommandWithArguments -ScriptBlock $ScriptBlock -ArgumentList $ArgumentList)"
```

There are downsides no matter how you slice it. Maybe we can come up with a way that has fewer downsides in the future. Going to try encoding the whole block next because we still have some quote-escaping stuff to deal with.